Narrow the resources for Terraform remote state access
It was pointed out that a wildcard was likely inappropriate as a scope for
tfstate files in the remote state. Except for the rare exception we would only
need access to `terraform.tfstate`. Since this project is _not_ one of those
exceptions we have limited the scope of the permission.

Co-authored-by: dav3r <david.redmin@trio.dhs.gov>
Co-authored-by: Shane Frasier <jeremy.frasier@trio.dhs.gov>