Add ownership controls to the S3 buckets
This implements a control to force all current and future objects in
the assessment images buckets (Production and Staging) to only be owned
by the bucket owner (the respective Images accounts). This disables
access control lists (ACLs) for bucket objects and forces the usage of
policies (such as IAM policies) to access objects. Since this is
already how we control access it makes sense to lock these buckets down
to make it a mandatory means of controlling access.

This change required updating the minimum AWS provider version used to
get support for this relatively new feature.